feat: Separating entities into two devices, with new entity naming conventions (#265)
BREAKING CHANGE: This release adds separates entities into two devices, and adds a default device for the boiler. With the addition of this new boiler entity, two old entities are removed (switch for turning on hot water boost and number representing boiler water temperature).

In addition to these changes, all entities are refactored to use new naming conventions for entities. This is not a breaking change for existing systems, but uninstalling and reinstalling the integration will rename some entities and their IDs.